# docs/swagger.yaml
basePath: /
definitions:
  models.RequestProxy:
    properties:
      headers:
        additionalProperties:
          type: string
        type: object
      method:
        type: string
      url:
        type: string
    type: object
  models.ResponseProxy:
    properties:
      headers:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      length:
        type: integer
      status:
        type: integer
    type: object
info:
  contact: {}
  description: API Server for Proxy Server App
  title: Proxy Server
  version: "1.0"
paths:
  /proxy:
    get:
      consumes:
      - application/json
      description: get all history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseProxy'
        "502":
          description: Internal server error
          schema:
            type: string
      summary: get all requests and responses
      tags:
      - proxy
    post:
      consumes:
      - application/json
      description: create request
      parameters:
      - description: request data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RequestProxy'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseProxy'
        "400":
          description: Invalid request body
          schema:
            type: string
        "502":
          description: Internal server error
          schema:
            type: string
      summary: request url
      tags:
      - proxy
  /proxy/{id}:
    get:
      consumes:
      - application/json
      description: get all history
      parameters:
      - description: request id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseProxy'
        "400":
          description: Invalid request body
          schema:
            type: string
        "502":
          description: Internal server error
          schema:
            type: string
      summary: get all requests and responses
      tags:
      - proxy
swagger: "2.0"